reorganization code
create Poly class is subclass of ofPolyline that has color value.
